Reorganize core persistence crate
