+++ nautilus_core/persistence/src/backend/session.rs
@@ -23,17 +23,12 @@ use datafusion::{
     prelude::*,
 };
 use futures::StreamExt;
-use nautilus_core::{ffi::cvec::CVec, python::to_pyruntime_err};
-use nautilus_model::data::{
-    bar::Bar, delta::OrderBookDelta, quote::QuoteTick, trade::TradeTick, Data, HasTsInit,
-};
-use pyo3::{prelude::*, types::PyCapsule};
+use nautilus_core::ffi::cvec::CVec;
+use nautilus_model::data::{Data, HasTsInit};
+use pyo3::prelude::*;
 
 use crate::{
-    arrow::{
-        DataStreamingError, DecodeDataFromRecordBatch, EncodeToRecordBatch, NautilusDataType,
-        WriteStream,
-    },
+    arrow::{DataStreamingError, DecodeDataFromRecordBatch, EncodeToRecordBatch, WriteStream},
     kmerge_batch::{EagerStream, ElementBatchIter, KMerge},
 };
 
@@ -61,12 +56,15 @@ pub type QueryResult = KMerge<EagerStream<std::vec::IntoIter<Data>>, Data, TsIni
 /// The session is used to register data sources and make queries on them. A
 /// query returns a Chunk of Arrow records. It is decoded and converted into
 /// a Vec of data by types that implement [`DecodeDataFromRecordBatch`].
-#[pyclass]
+#[cfg_attr(
+    feature = "python",
+    pyclass(module = "nautilus_trader.core.nautilus_pyo3.persistence")
+)]
 pub struct DataBackendSession {
     session_ctx: SessionContext,
     batch_streams: Vec<EagerStream<IntoIter<Data>>>,
-    chunk_size: usize,
-    runtime: Arc<tokio::runtime::Runtime>,
+    pub chunk_size: usize,
+    pub runtime: Arc<tokio::runtime::Runtime>,
 }
 
 impl DataBackendSession {
@@ -176,110 +174,23 @@ impl DataBackendSession {
 // Note: Intended to be used on a single python thread
 unsafe impl Send for DataBackendSession {}
 
-////////////////////////////////////////////////////////////////////////////////
-// Python API
-////////////////////////////////////////////////////////////////////////////////
-#[cfg(feature = "python")]
-#[pymethods]
-impl DataBackendSession {
-    #[new]
-    #[pyo3(signature=(chunk_size=5_000))]
-    fn new_session(chunk_size: usize) -> Self {
-        Self::new(chunk_size)
-    }
-
-    /// Query a file for its records. the caller must specify `T` to indicate
-    /// the kind of data expected from this query.
-    ///
-    /// table_name: Logical table_name assigned to this file. Queries to this file should address the
-    /// file by its table name.
-    /// file_path: Path to file
-    /// sql_query: A custom sql query to retrieve records from file. If no query is provided a default
-    /// query "SELECT * FROM <table_name>" is run.
-    ///
-    /// # Safety
-    /// The file data must be ordered by the ts_init in ascending order for this
-    /// to work correctly.
-    #[pyo3(name = "add_file")]
-    fn add_file_py(
-        mut slf: PyRefMut<'_, Self>,
-        data_type: NautilusDataType,
-        table_name: &str,
-        file_path: &str,
-        sql_query: Option<&str>,
-    ) -> PyResult<()> {
-        let _guard = slf.runtime.enter();
-
-        match data_type {
-            NautilusDataType::OrderBookDelta => slf
-                .add_file::<OrderBookDelta>(table_name, file_path, sql_query)
-                .map_err(to_pyruntime_err),
-            NautilusDataType::QuoteTick => slf
-                .add_file::<QuoteTick>(table_name, file_path, sql_query)
-                .map_err(to_pyruntime_err),
-            NautilusDataType::TradeTick => slf
-                .add_file::<TradeTick>(table_name, file_path, sql_query)
-                .map_err(to_pyruntime_err),
-            NautilusDataType::Bar => slf
-                .add_file::<Bar>(table_name, file_path, sql_query)
-                .map_err(to_pyruntime_err),
-        }
-    }
-
-    fn to_query_result(mut slf: PyRefMut<'_, Self>) -> DataQueryResult {
-        let query_result = slf.get_query_result();
-        DataQueryResult::new(query_result, slf.chunk_size)
-    }
-}
-
-#[pyclass]
+#[cfg_attr(
+    feature = "python",
+    pyclass(module = "nautilus_trader.core.nautilus_pyo3.persistence")
+)]
 pub struct DataQueryResult {
-    result: QueryResult,
     chunk: Option<CVec>,
-    acc: Vec<Data>,
-    size: usize,
-}
-
-#[cfg(feature = "python")]
-#[pymethods]
-impl DataQueryResult {
-    /// The reader implements an iterator.
-    fn __iter__(slf: PyRef<'_, Self>) -> PyRef<'_, Self> {
-        slf
-    }
-
-    /// Each iteration returns a chunk of values read from the parquet file.
-    fn __next__(mut slf: PyRefMut<'_, Self>) -> PyResult<Option<PyObject>> {
-        slf.drop_chunk();
-
-        for _ in 0..slf.size {
-            match slf.result.next() {
-                Some(item) => slf.acc.push(item),
-                None => break,
-            }
-        }
-
-        let mut acc: Vec<Data> = Vec::new();
-        std::mem::swap(&mut acc, &mut slf.acc);
-
-        if !acc.is_empty() {
-            let cvec = acc.into();
-            Python::with_gil(|py| match PyCapsule::new::<CVec>(py, cvec, None) {
-                Ok(capsule) => Ok(Some(capsule.into_py(py))),
-                Err(err) => Err(to_pyruntime_err(err)),
-            })
-        } else {
-            Ok(None)
-        }
-    }
+    pub result: QueryResult,
+    pub acc: Vec<Data>,
+    pub size: usize,
 }
 
 impl DataQueryResult {
     #[must_use]
     pub fn new(result: QueryResult, size: usize) -> Self {
         Self {
-            result,
             chunk: None,
+            result,
             acc: Vec::new(),
             size,
         }
@@ -288,7 +199,7 @@ impl DataQueryResult {
     /// Chunks generated by iteration must be dropped after use, otherwise
     /// it will leak memory. Current chunk is held by the reader,
     /// drop if exists and reset the field.
-    fn drop_chunk(&mut self) {
+    pub fn drop_chunk(&mut self) {
         if let Some(CVec { ptr, len, cap }) = self.chunk.take() {
             let data: Vec<Data> =
                 unsafe { Vec::from_raw_parts(ptr.cast::<nautilus_model::data::Data>(), len, cap) };
