+++ nautilus_core/persistence/src/lib.rs
@@ -18,18 +18,5 @@ pub mod backend;
 mod kmerge_batch;
 pub mod wranglers;
 
-use pyo3::prelude::*;
-
-/// Loaded as nautilus_pyo3.persistence
-#[pymodule]
-pub fn persistence(_: Python<'_>, m: &PyModule) -> PyResult<()> {
-    m.add_class::<arrow::NautilusDataType>()?;
-    m.add_class::<backend::session::DataBackendSession>()?;
-    m.add_class::<backend::session::DataQueryResult>()?;
-    m.add_class::<backend::transformer::DataTransformer>()?;
-    m.add_class::<wranglers::bar::BarDataWrangler>()?;
-    m.add_class::<wranglers::delta::OrderBookDeltaDataWrangler>()?;
-    m.add_class::<wranglers::quote::QuoteTickDataWrangler>()?;
-    m.add_class::<wranglers::trade::TradeTickDataWrangler>()?;
-    Ok(())
-}
+#[cfg(feature = "python")]
+pub mod python;
