+++ nautilus_core/persistence/src/python/backend/mod.rs
@@ -14,3 +14,4 @@
 // -------------------------------------------------------------------------------------------------
 
 pub mod session;
+pub mod transformer;
