+++ nautilus_core/persistence/src/python/mod.rs
@@ -13,4 +13,20 @@
 //  limitations under the License.
 // -------------------------------------------------------------------------------------------------
 
-pub mod session;
+use pyo3::prelude::*;
+
+mod backend;
+
+/// Loaded as nautilus_pyo3.persistence
+#[pymodule]
+pub fn persistence(_: Python<'_>, m: &PyModule) -> PyResult<()> {
+    m.add_class::<crate::arrow::NautilusDataType>()?;
+    m.add_class::<crate::backend::session::DataBackendSession>()?;
+    m.add_class::<crate::backend::session::DataQueryResult>()?;
+    m.add_class::<backend::transformer::DataTransformer>()?;
+    m.add_class::<crate::wranglers::bar::BarDataWrangler>()?;
+    m.add_class::<crate::wranglers::delta::OrderBookDeltaDataWrangler>()?;
+    m.add_class::<crate::wranglers::quote::QuoteTickDataWrangler>()?;
+    m.add_class::<crate::wranglers::trade::TradeTickDataWrangler>()?;
+    Ok(())
+}
